Further fix for BZ 57172 - followup to r1640087 (r1640084 in trunk). Be consistent with "webappClassLoader.stopped" message handling and throw an ISE from findResourceInternal
In all call chains (findClass() / findResource() public methods) there is a checkStateForResourceLoading()/checkStateForClassLoading() check at the start of that public method, so it is unlikely to hit an ISE here unless there is some kind of race.  Throwing an ISE is a more consistent behaviour, and calling a common method is better than a custom rarely used if() block.

git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1640410 13f79535-47bb-0310-9956-ffa450edef68