Supporting "+" and "-" syntax in SSL protocol (OpenSSL) only.
This allows settings like "all-TLSv1" (or if you
prefer "all,-TLSv1"). The preferred way to add
protocols now is to use an explicit "+" prefix.
If you omit the "+" (except for the first token
in the list) you get a warning. The list is always
build starting with an empty list, even if the
first token has a "+" prefix.

This is similar to the mod_ssl behavior. It differs
in the handling of a token without "+" or "-":
mod_ssl would reset the list, we just add to the list,
both for compatibility reasons.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1681779 13f79535-47bb-0310-9956-ffa450edef68