Fix class loader decision on the delegation for class loading and resource lookup and make it faster too.
This is the first in a series of changes in
WebappClassLoaderBase.filter().

Filtering for resources without a dot in their
name is broken due to checking for a dot.

The check was needed in older versions of
that code which did prefix comparisons.
The current code using regexps does no
longer need the check. For any class loading
or resource lookup with a dot in the name
(mostly always the case) it will only slow
down the filtering.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1726632 13f79535-47bb-0310-9956-ffa450edef68